fix(list): return `unsigned int` from dl_list_len
Return `unsigned int` from dl_list_len instead of `ssize_t`.

This is what the hostapd source code returns,
see https://w1.fi/cgit/hostap/tree/src/utils/list.h?id=0ccb3b6cf25e0c432de092a5298be96accd18c68#n54,
and is what this function will return once
https://github.com/nqminds/edgesec/pull/265 is merged.

Fixes missing `ssize_t` when using `list.h` when using standard ISO C.

Issues:
  - Some functions convert the result of `dl_list_len` to a `ssize_t`.
    On 64-bit Linux, this is fine, since `unsigned int` is 32-bit,
    and `ssize_t` is 64-bit.
    However, on 32-bit Linux, this is an issue, since the
    `unsigned int` is 32-bit, and the `ssize_t` is 32-bit, so
    the conversion is undefined if it overflows.

+++ src/utils/list.h
@@ -56,9 +56,9 @@ static inline int dl_list_empty(struct dl_list *list) {
   return list->next == list;
 }
 
-static inline ssize_t dl_list_len(struct dl_list *list) {
+static inline unsigned int dl_list_len(struct dl_list *list) {
   struct dl_list *item;
-  ssize_t count = 0;
+  int count = 0;
   for (item = list->next; item != list; item = item->next)
     count++;
   return count;
